Rearranging render macros
