+++ src/module/macros.ts
@@ -10,7 +10,6 @@ import type {
   LancerReactionMacroData,
   LancerStatMacroData,
   LancerTalentMacroData,
-  LancerTechMacroData,
   LancerTextMacroData,
 } from "./interfaces";
 // Import JSON data
@@ -36,27 +35,20 @@ import { StabOptions1, StabOptions2 } from "./enums";
 import { applyCollapseListeners, uuid4 } from "./helpers/collapse";
 import { checkForHit } from "./helpers/automation/targeting";
 import type { AccDiffData, AccDiffDataSerialized, RollModifier } from "./helpers/acc_diff";
-import { is_limited, is_overkill, is_tagged } from "machine-mind/dist/funcs";
+import { is_limited, is_overkill } from "machine-mind/dist/funcs";
 import type { LancerToken } from "./token";
 import { is_loading, is_self_heat } from "machine-mind/dist/classes/mech/EquipUtil";
 import { getAutomationOptions } from "./settings";
 
-import {
-  getMacroSpeaker,
-  encodeMacroData,
-  encodedMacroWhitelist,
-} from "./macros/util"
+import { getMacroSpeaker, encodeMacroData, encodedMacroWhitelist } from "./macros/util"
+import { renderMacroTemplate, renderMacroHTML } from "./macros/render"
+import { prepareTechMacro, rollTechMacro } from "./macros/tech"
+import { prepareTextMacro, rollTextMacro } from "./macros/text"
+
 export { encodeMacroData, runEncodedMacro } from "./macros/util"
+export { renderMacroTemplate, renderMacroHTML } from "./macros/render"
 export { prepareActivationMacro } from "./macros/activation"
-import {
-  prepareTechMacro,
-  rollTechMacro,
-} from "./macros/tech"
 export { prepareTechMacro } from "./macros/tech"
-import {
-  prepareTextMacro,
-  rollTextMacro,
-} from "./macros/text"
 export { prepareTextMacro } from "./macros/text"
 
 const lp = LANCER.log_prefix;
@@ -320,55 +312,6 @@ export async function prepareItemMacro(a: string, i: string, options?: any) {
   }
 
   applyCollapseListeners();
-}
-
-/**
- *
- */
-// TODO: Indexed types for templates
-export async function renderMacroTemplate(actor: LancerActor | undefined, template: string, templateData: any) {
-  const cardUUID = uuid4();
-  templateData._uuid = cardUUID;
-
-  const html = await renderTemplate(template, templateData);
-
-  // Schlorp up all the rolls into a mega-roll so DSN sees the stuff to throw
-  // on screen
-  const aggregate: Roll[] = [];
-  if (templateData.roll) {
-    aggregate.push(templateData.roll);
-  }
-  if (templateData.attacks) {
-    aggregate.push(...templateData.attacks.map((a: { roll: Roll }) => a.roll));
-  }
-  if (templateData.crit_damages) {
-    aggregate.push(...templateData.crit_damages.map((d: { roll: Roll }) => d.roll));
-  } else if (templateData.damages) {
-    aggregate.push(...templateData.damages.map((d: { roll: Roll }) => d.roll));
-  }
-  const roll = Roll.fromTerms([PoolTerm.fromRolls(aggregate)]);
-
-  return renderMacroHTML(actor, html, roll);
-}
-
-export async function renderMacroHTML(actor: LancerActor | undefined, html: HTMLElement | string, roll?: Roll) {
-  const rollMode = game.settings.get("core", "rollMode");
-  const chat_data = {
-    user: game.user,
-    type: roll ? CONST.CHAT_MESSAGE_TYPES.ROLL : CONST.CHAT_MESSAGE_TYPES.IC,
-    roll: roll,
-    speaker: {
-      actor: actor,
-      token: actor?.token,
-      alias: !!actor?.token ? actor.token.name : null,
-    },
-    content: html,
-    whisper: rollMode !== "roll" ? ChatMessage.getWhisperRecipients("GM").filter(u => u.active) : undefined,
-  };
-  // @ts-ignore This is fine
-  const cm = await ChatMessage.create(chat_data);
-  cm?.render();
-  return Promise.resolve();
 }
 
 /** TODO: Remove if not needed
